Elements. Remove PrefixElementImpl.
For now, we have a not nice looking way to write / read Element(s). Most
of the elements are stored as their first fragment, and then asked for
elements. I start with PrefixElementImpl2 to stop doing this, and switch
to writing actual element reference. When we are done with all elements,
it will become nice again.

Change-Id: Ia18c2ab516ae06dc0da0b1ab022f217bf9a7400d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/431702
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>