Add a test for hasPublishedDocs
This causes the code that parses the `messages.yaml` files in either
the analyzer or linter packages so that it throw an exception if any
code fails to have an explicit `hasPublishedDocs` key.

We decided to require the field because it makes it easier to identify
codes for which documentation has not yet been published. Prior to this
the search looked for codes missing this key _or_ codes where this key
had a value of `false`, but now only the last needs to be searched for
which can be done using a simple textual search.

Change-Id: I1aa59eb02a628fae8cb79c5390b59c787218fe15
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/431761
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>