Normalize the diagnostic pages
This normalizes the names of the classes to match the label displayed to
the user (to make it easier to find the class that needs to be updated).

It also normalizes capitalization in titles and the formation of the
page ids.

This is a precursor to splitting the page classes out into separate files.

Change-Id: I571bfaf23057f2366274e4b853412c8756151f65
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/432121
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>