[cfe] Normalize handling of extension constructor
This normalizes the handling of extension constructor to make it similar
to constructors of other declarations. Extension constructors are
erroneous but were partially present in some parts of the processing of
builders. This change handles them fully, similar to extension type
constructors, and just avoids emitting them. It might be that this some
what prepares for the static extension feature as well.

Change-Id: Icee027a2edb45c1b42b1f92d90150d6827c45a80
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/432981
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>