[analysis_server] Fix compilation of server when using custom SDK
When running analysis server tests from source, we now (since a recent change) pre-compile the server to avoid compilation on every test.

If the first test run was `test\integration\analysis\get_errors_non_standard_sdk_test.dart` we would try to compile with the non-standard (and incomplete) SDK created by the test, and fail because of https://github.com/dart-lang/sdk/issues/52752. This changes to always use the real SDK that we're running with for the compilation instead.

Change-Id: I1028c24ecb18f9caa465b579866d07a72c1dc02b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/433822
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>