Add test to check that dot shorthands do not have an obvious type
This PR adds a test to confirm that the lint
`omit_obvious_local_variable_types` considers a dot shorthand to be an
expression whose type is not obvious. This is crucial because the dot
shorthand would no longer work if the context type is turned into `_`
because the type annotation is removed. Similarly for the lint
`omit_obvious_property_types`.

Fixes https://github.com/dart-lang/sdk/issues/60910
Change-Id: Ie31026286eb496c0cd76c7bb5ca64e04340b4aaa
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/434660
Commit-Queue: Phil Quitslund <pquitslund@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>