analyzer: Move the DevExp style guide doc to analysis_server
This is awkward as there are some packages which are co-developed like
analyzer_testing and analyzer_utilities. But broadly speaking, this
style guide is used for the "analysis-related" packages that the Dash
Developer Experience team maintains. So we can move it to live with the
other docs in pkg/analysis_server.

Change-Id: Ic94cfe1d16288ec707c96a42c7a18788e9f8b7d0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/435220
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>