[dart2wasm] Fix remaining issues around [Symbol] and privacy
Symbol comparison & hashing has to take library privacy into account.

We refactor the compiler to move symbol related logic to a new
`pkg/dart2wasm/lib/symbols.dart:Symbols`. It will be responsible for
making symbol constants, mangling and minification.

In there we encapsulate the name mangling required to make library
privacy work (similar to how the VM does it).

We also clean up the runtime implementation (which was a copy of
the VM) and make the unmangling be based purely on removing trailing
`...@<number>`.

Fixes the following tests

  * co19/LibTest/core/Symbol/Symbol_A03_t01
  * co19/LibTest/core/Symbol/Symbol_A03_t02
  * language/private/private_test

Closes https://github.com/dart-lang/sdk/issues/50791

Change-Id: If1bb19f3d8868662531162a0458754007017df33
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/435140
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>