analyzer: Privatize ParsedFileState, follow-on work
Work towards https://github.com/dart-lang/sdk/issues/60635

I set out to rename `ParsedFileState.errors` and was concerned it might
be a breaking change, but it turned out that the class could be made
private. As well as ParsedFileStateCache.

But then `getParsed` would return a private type. That could be fixed
by making `getParsed` private.

Also the `parsedFileStateCache` getter seemed suspicious. Turned out it
was only used outside the library in order to clear the cache. So I make
the field private (at which point analyzer told me it should be
final), and added a clearing method.

Change-Id: Iaf30a9cc2a0cfb5c9168db243b7fb2f6c624d8e1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/433264
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>