[cfe] Create SourceConstructorBuilder from ConstructorDeclaration
This is a step toward supporting synthetic declaration to be created
together with fragment-based declarations.

A ConstructorReferences object is added to split the references for
constructors and their tear-off from the creation of the AST nodes,
enabling moving the creation of AST nodes for constructors to
buildOutlineNodes.

Change-Id: I0af9a6a51107092026c2dfd2031093f74c02e101
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/435122
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>