[analysis_server] Dot shorthands: Update prefer_const_constructors lint.
Now that `DotShorthandsConstructorInvocation`'s `canBeConst` is updated, we can also update the `prefer_const_constructors` lint.

We skip the lint when we have the `@literal` metadata on the constructor since it's already reporting a warning here: https://dart-review.googlesource.com/c/sdk/+/434881. Similar to instance creation expressions, we also skip reporting a lint if we're instantiating the `Object` class.

Otherwise, if the constructor we're calling with the shorthand is const and we can const-ify it, we report this lint.

Added unit tests.

Fixes: https://github.com/dart-lang/sdk/issues/60911
Bug: https://github.com/dart-lang/sdk/issues/60893
Change-Id: Ib4bcba629f60e0dbff110080918adfacfb1ba5d2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/434981
Commit-Queue: Kallen Tu <kallentu@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>