[dart2wasm] Fix reporting measurements in WasmJSInterop benchmarks
benchmark_harness's `measureFor` returns in microseconds, but we
currently show the numbers with the unit "ns".

Convert the microseconds reported by benchmark_harness to nanoseconds
when reporting.

We could also report in microseconds, but the numbers would be very
small, with a few zeros before fractional digits. So keep reporting in
nanoseconds.

Change-Id: I189190b2139c8e2d39b8bef84159585ab0967980
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/436904
Commit-Queue: Ömer Ağacan <omersa@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>